some changes

+++ routes/listing.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const listing = require("../models/listing.js");
-const {isLogedIn} = require("../middleware.js");
+const {isLogedIn, isowner} = require("../middleware.js");
 
 
 
@@ -19,7 +19,7 @@ router.get("/new",isLogedIn,(req,res) =>{
 router.get("/:id",async(req,res) =>{
     console.log(req.user);
         let {id} = req.params;
-        let showListing =await listing.findById(id).populate('reviews');
+        let showListing =await listing.findById(id).populate({path: "reviews", populate:{path: "author"}}).populate("owner");
         if(!showListing){
         req.flash("fail","Listing doesn't exist");
             res.redirect("/listings");
@@ -31,6 +31,8 @@ router.get("/:id",async(req,res) =>{
 
 router.post("/",isLogedIn,async(req,res) =>{
     let newlisting = await new listing(req.body.listing);
+    console.log(res.locals.currUser);
+    newlisting.owner = res.locals.currUser;
     newlisting.save();
     req.flash("success","listing is added");
     res.redirect("/listings");
@@ -43,7 +45,13 @@ router.get("/edit/:id",isLogedIn,async(req,res) =>{
     res.redirect("/listings");
     }
     else{
+        if(!newlisting.owner._id.equals(res.locals.currUser._id)){
+            req.flash("fail","You do not change this listing information");
+            return res.redirect(`/listings/${id}`);
+            }
         res.render("./listing/edit.ejs",{newlisting});
+        // console.log("newlisting.owner: ",newlisting.owner);
+        // console.log("res.locals.currUser",res.locals.currUser);
     }
 });
 
@@ -57,6 +65,10 @@ router.put("/edit/:id",isLogedIn,async(req,res) =>{
 
 router.delete("/delete/:id",isLogedIn,async(req,res) =>{
     let {id} = req.params;
+    if(!await listing.owner.equals(res.locals.currUser._id)){
+        req.flash("fail","You do not change this listing information");
+        return res.redirect(`/listings/${id}`);
+        }
     await listing.findByIdAndDelete(id);
     req.flash("success","listing is deleted successfully");
     res.redirect("/listings");
